+++ packages/nibijs/src/msg/staking.ts
@@ -1,4 +1,4 @@
-import { EncodeObject, GeneratedType, Registry } from "@cosmjs/proto-signing"
+import { Registry } from "@cosmjs/proto-signing"
 import * as pb from "@nibiruchain/protojs/dist/cosmos/staking/v1beta1/tx"
 import { MsgTypeUrls, TxMessage } from "./types"