+++ packages/protojs/src/spot/v1/genesis.ts
@@ -1,32 +1,23 @@
 /* eslint-disable */
 import Long from "long";
 import _m0 from "protobufjs/minimal";
-import { Params, PostedPrice } from "./state";
+import { Params } from "./params";
 
-export const protobufPackage = "nibiru.pricefeed.v1";
+export const protobufPackage = "nibiru.spot.v1";
 
-/** GenesisState defines the pricefeed module's genesis state. */
+/** GenesisState defines the spot module's genesis state. */
 export interface GenesisState {
-  /** params defines all the paramaters of the module. */
   params?: Params;
-  postedPrices: PostedPrice[];
-  genesisOracles: string[];
 }
 
 function createBaseGenesisState(): GenesisState {
-  return { params: undefined, postedPrices: [], genesisOracles: [] };
+  return { params: undefined };
 }
 
 export const GenesisState = {
   encode(message: GenesisState, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
     if (message.params !== undefined) {
       Params.encode(message.params, writer.uint32(10).fork()).ldelim();
-    }
-    for (const v of message.postedPrices) {
-      PostedPrice.encode(v!, writer.uint32(18).fork()).ldelim();
-    }
-    for (const v of message.genesisOracles) {
-      writer.uint32(26).string(v!);
     }
     return writer;
   },
@@ -41,12 +32,6 @@ export const GenesisState = {
         case 1:
           message.params = Params.decode(reader, reader.uint32());
           break;
-        case 2:
-          message.postedPrices.push(PostedPrice.decode(reader, reader.uint32()));
-          break;
-        case 3:
-          message.genesisOracles.push(reader.string());
-          break;
         default:
           reader.skipType(tag & 7);
           break;
@@ -56,28 +41,12 @@ export const GenesisState = {
   },
 
   fromJSON(object: any): GenesisState {
-    return {
-      params: isSet(object.params) ? Params.fromJSON(object.params) : undefined,
-      postedPrices: Array.isArray(object?.postedPrices)
-        ? object.postedPrices.map((e: any) => PostedPrice.fromJSON(e))
-        : [],
-      genesisOracles: Array.isArray(object?.genesisOracles) ? object.genesisOracles.map((e: any) => String(e)) : [],
-    };
+    return { params: isSet(object.params) ? Params.fromJSON(object.params) : undefined };
   },
 
   toJSON(message: GenesisState): unknown {
     const obj: any = {};
     message.params !== undefined && (obj.params = message.params ? Params.toJSON(message.params) : undefined);
-    if (message.postedPrices) {
-      obj.postedPrices = message.postedPrices.map((e) => e ? PostedPrice.toJSON(e) : undefined);
-    } else {
-      obj.postedPrices = [];
-    }
-    if (message.genesisOracles) {
-      obj.genesisOracles = message.genesisOracles.map((e) => e);
-    } else {
-      obj.genesisOracles = [];
-    }
     return obj;
   },
 
@@ -86,8 +55,6 @@ export const GenesisState = {
     message.params = (object.params !== undefined && object.params !== null)
       ? Params.fromPartial(object.params)
       : undefined;
-    message.postedPrices = object.postedPrices?.map((e) => PostedPrice.fromPartial(e)) || [];
-    message.genesisOracles = object.genesisOracles?.map((e) => e) || [];
     return message;
   },
 };