+++ packages/nibijs/CHANGELOG.md
@@ -7,6 +7,20 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 * .
 
+## v0.19.2
+
+* chore: Raise minimum gas limit for transactions with low values for the simulated "gas used".
+
+## v0.19.1 
+
+* feat: Transactions for the distribution module 
+* feat: Queries for the distribution module 
+* sync: Update to protojs v0.19.0
+
+## v0.8.6
+
+* feat: Add staking query extension
+
 ## v0.8.0 
 
 * sync: Update to nibiru v0.16.0-beta.1.